Move text layout to KDraw
